new matlab folder forward_turn is working now with constant speed

+++ API_Robot/motor.h
@@ -10,7 +10,7 @@ class Motor
 public:
 	Motor();
 	Motor(int motor);
-	void setVelocity(float velocity);
+	void setVelocity(double velocity);
 	//setVoltage();
 	//void enableMotors();
 
@@ -18,10 +18,10 @@ private:
 	int _motor;
 	int _dir_pin;
 	int _pwm_pin;
-	float _velocity;
+	double _velocity;
 	int _pwm;
 };
 
-int vel2pwm(float velocity);
+int vel2pwm(double velocity);
 
 #endif
